feat: show chats messages
